+++ src/main/java/uk/gov/hmcts/reform/professionalapi/domain/AccessType.java
@@ -25,15 +25,4 @@ public class AccessType {
         this.enabled = enabled;
     }
 
-    public static AccessType fromUserConfiguredAccess(UserConfiguredAccess userConfiguredAccess) {
-        AccessType accessType = new AccessType();
-        accessType.setAccessTypeId(userConfiguredAccess.getUserConfiguredAccessId().getAccessTypeId());
-        accessType.setOrganisationProfileId(userConfiguredAccess.getUserConfiguredAccessId()
-                .getOrganisationProfileId());
-        accessType.setJurisdictionId(userConfiguredAccess.getUserConfiguredAccessId().getJurisdictionId());
-        accessType.setEnabled(userConfiguredAccess.getEnabled());
-
-        return accessType;
-    }
-
 }
