+++ src/test/java/uk/gov/hmcts/reform/professionalapi/service/impl/ProfessionalUserServiceImplTest.java
@@ -53,7 +53,6 @@ import uk.gov.hmcts.reform.professionalapi.util.RefDataUtil;
 
 import java.nio.charset.Charset;
 import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.List;
@@ -75,6 +74,7 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 import static org.springframework.http.MediaType.APPLICATION_JSON;
 import static uk.gov.hmcts.reform.professionalapi.controller.constants.ProfessionalApiConstants.ERROR_MESSAGE_UP_FAILED;
+import static uk.gov.hmcts.reform.professionalapi.controller.constants.ProfessionalApiConstants.ISO_DATE_TIME_FORMATTER;
 import static uk.gov.hmcts.reform.professionalapi.controller.constants.ProfessionalApiConstants.LENGTH_OF_ORGANISATION_IDENTIFIER;
 import static uk.gov.hmcts.reform.professionalapi.generator.ProfessionalApiGenerator.generateUniqueAlphanumericId;
 
@@ -1196,8 +1196,7 @@ class ProfessionalUserServiceImplTest {
         when(professionalUserPage.getContent()).thenReturn(professionalUserList);
 
         LocalDateTime currentDateTime = LocalDateTime.of(2023,12,6,13,36,25);
-        DateTimeFormatter formatter = DateTimeFormatter.ISO_DATE_TIME;
-        String since = currentDateTime.format(formatter);
+        String since = currentDateTime.format(ISO_DATE_TIME_FORMATTER);
 
         ResponseEntity<Object> responseEntity = professionalUserService.findRefreshUsers(since, pageableMock);
